+++ ChangeLog.md
@@ -21,6 +21,8 @@ Released 2017/01/11
 * Drop support for compute 2.0 GPU devices
 * Support cusolver with CUDA 8.0
 
+* Add `hoomd.hdf5.log` to log quantities in hdf5 format. Matrix quantities can be logged.
+
 ## v2.1.1
 
 Released 2016/10/23